backend/requirements.txt
xmltodict
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-cors-headers==3.0.2
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.2
idna==2.8
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
